// src/wasm/pkg/anvil_wasm.d.ts
export enum AlphaBlurMode {
  /**
   * Skip alpha channel (preserve original alpha values)
   */
  Skip = 0,
  /**
   * Apply blur to alpha channel as well
   */
  Blur = 1,
}
export enum AntialiasMode {
  Nearest = 0,
  Bilinear = 1,
  Bicubic = 2,
}
export enum DitheringMode {
  /**
   * Random dithering (white noise)
   */
  Random = 0,
  /**
   * Floyd-Steinberg error diffusion
   */
  ErrorDiffusion = 1,
  /**
   * Ordered dithering using Bayer matrix
   */
  Ordered = 2,
}
// ...
export class RgbaBuffer {
  free(): void;
  [Symbol.dispose](): void;
  dustRemoval(max_size: number, alpha_threshold: number): void;
  gaussianBlur(radius: number, alpha_mode: AlphaBlurMode): void;
  brightnessAndContrast(brightness: number, contrast: number): void;
  invert(): void;
  dithering(mode: DitheringMode, levels: number, strength: number): void;
  grayscale(): void;
  posterize(levels: number): void;
  blitFromRaw(source: Uint8Array, source_width: number, source_height: number, offset_x: number, offset_y: number, scale_x: number, scale_y: number, rotate_deg: number, antialias_mode: AntialiasMode, flip_x: boolean, flip_y: boolean): void;
  cropWithMask(mask: Uint8Array, mask_width: number, mask_height: number, mask_offset_x: number, mask_offset_y: number): Uint8Array;
  sliceWithMask(mask: Uint8Array, mask_width: number, mask_height: number, mask_offset_x: number, mask_offset_y: number): Uint8Array;
  blitFromBuffer(source: RgbaBuffer, offset_x: number, offset_y: number, scale_x: number, scale_y: number, rotate_deg: number, antialias_mode: AntialiasMode, flip_x: boolean, flip_y: boolean): void;
  writeRect(rect_x: number, rect_y: number, rect_width: number, rect_height: number, data: Uint8Array): boolean;
  writePixels(coords: Uint32Array, colors: Uint8Array): boolean;
  readRect(rect_x: number, rect_y: number, rect_width: number, rect_height: number): Uint8Array;
  exportPng(): Uint8Array;
  exportWebp(): Uint8Array;
  resizeWithOrigins(new_width: number, new_height: number, src_origin_x: number, src_origin_y: number, dest_origin_x: number, dest_origin_y: number): void;
  resize(new_width: number, new_height: number): void;
  overwriteWith(raw: Uint8Array, width: number, height: number): boolean;
  get(x: number, y: number): Array<any>;
  len(): number;
  constructor(width: number, height: number);
  ptr(): number;
  set(x: number, y: number, r: number, g: number, b: number, a: number): boolean;
  clone(): RgbaBuffer;
  width(): number;
  height(): number;
  static fromRaw(width: number, height: number, buf: Uint8Array): RgbaBuffer;
  is_empty(): boolean;
  data(): Uint8ClampedArray;
  static fromWebp(width: number, height: number, webp_buf: Uint8Array): RgbaBuffer;
  isInBounds(x: number, y: number): boolean;
  indexGet(idx: number): Array<any>;
  indexSet(idx: number, r: number, g: number, b: number, a: number): boolean;
  floodFill(start_x: number, start_y: number, fill_color_r: number, fill_color_g: number, fill_color_b: number, fill_color_a: number, threshold: number): boolean;
  fillAllCodes(fill_code: number): boolean;
  fillMaskArea(mask: Uint8Array, fill_color_r: number, fill_color_g: number, fill_color_b: number, fill_color_a: number): boolean;
  fillAllPixels(fill_color_r: number, fill_color_g: number, fill_color_b: number, fill_color_a: number): boolean;
  floodFillWithMask(start_x: number, start_y: number, fill_color_r: number, fill_color_g: number, fill_color_b: number, fill_color_a: number, threshold: number, selection_mask: Uint8Array, limit_mode: string): boolean;
  importPng(png_buffer: Uint8Array, width: number, height: number): boolean;
  importRaw(raw: Uint8Array, width: number, height: number): boolean;
  importWebp(webp_buffer: Uint8Array, width: number, height: number): boolean;
}